*** empty log message ***
git-svn-id: http://sqlworkbench.mgm-tp.com/sqlworkbench/trunk/sqlworkbench@344 6c36a7f0-0884-4472-89c7-80b7fa7b7f28

+++ src/workbench/gui/dbobjects/ProgressPanel.java
@@ -81,11 +81,11 @@ public class ProgressPanel
 	{
 	}
 
-	public void setCurrentObject(String object, int number, int totalObjects)
+	public void setCurrentObject(String object, long number, long totalObjects)
 	{
 	}
 
-	public void setCurrentRow(int currentRow, int totalRows)
+	public void setCurrentRow(long currentRow, long totalRows)
 	{
 		if (currentRow > -1) this.rowInfo.setText(Long.toString(currentRow));
 	}
